requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
flake8
pytest==7.4.4
httpx==0.26.0
python-multipart==0.0.6
email-validator==2.1.0.post1
cryptography==41.0.7
pandas
numpy
alembic==1.13.1
kagglehub[pandas-datasets]
rich==13.7.0
backoff